pyproject.toml

[project]
name = "redbrick-sdk"
description = "RedBrick platform Python SDK!"
readme = "README.md"
license.file = "LICENSE"
dynamic = ["version"]
requires-python = ">=3.9,<3.14"
keywords = ["redbrick"]
dependencies = [
    "aiohttp<4,>=3.10.5",
    "dicom2nifti<3",
    "inquirerpy<1",
    "natsort<9,>=8.0.2",
    "nest-asyncio<2,>=1.5.4",
    "nibabel<6,>=3.2.2",
    "numpy<3,>=1.15",
    "packaging<25",
    "Pillow<11,>=9.0.1",
    "pydicom<3,>=2.4",
    "pydicom-seg<1,>=0.4.1",
    "python-dateutil<3",
    "requests<3,>=2.23.0",
    "Rich<14",
    "rt-utils<2",
    "shtab<2",
    "SimpleITK<3",
    "tenacity<10",
    "tqdm<5",
    "typing-extensions<5",
]
classifiers = [
  "Programming Language :: Python"
]

[project.scripts]
redbrick = "redbrick.cli:cli_main"

[project.urls]
Homepage = "https://www.redbrickai.com"
Documentation = "https://sdk.redbrickai.com"
Changelog = "https://github.com/redbrick-ai/redbrick-sdk/releases"

[project.optional-dependencies]
dev = [
    "black<=24.10.0",
    "build<=1.2.2.post1",
    "flake8<=7.1.1",
    "gputil<=1.4.0",
    "mypy<=1.13.0",
    "mypy-extensions<=1.0.0",
    "psutil<=6.1.0",
    "pycodestyle<=2.12.1",
    "pydicom<4,>=2.4",
    "pydocstyle<=6.3.0",
    "pylint<=3.3.2",
    "pytest<=8.3.4",
    "pytest-asyncio<=0.25.0",
    "pytest-cov<=6.0.0",
    "pytest-randomly<=3.16.0",
    "pytest-xdist<=3.6.1",
    "pytest-xdist[psutil]<=3.6.1",
    "setuptools<=75.6.0",
    "wheel<=0.45.1",
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.setuptools.dynamic]
version = {attr = "redbrick.__version__"}
readme = {file = ["README.rst"]}

[tool.setuptools.packages.find]
where = ["."]
include = ["redbrick*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
redbrick = ["py.typed"]
